Added Property Resolving On Return
Now, if you add an object or array of objects to the response method, it will resolve either automatic properties as key and values in the response array. Also can have a class implement the ResponseableInterface to customize that class's return format.  Also got rid of all static methods and unnecessary facades and service providers.